+++ scripts/sp_UpdateMembersCapitalAccount.sql
@@ -11,7 +11,7 @@ END
 
 GO
 
-CREATE PROCEDURE [dbo].[sp_UpdateMembersCapitalAccount](@ValuationDate as DATETIME, @Member as nvarchar(256), @Units as float, @Account as INT) AS
+CREATE PROCEDURE [dbo].[sp_UpdateMembersCapitalAccount](@ValuationDate as DATETIME, @Member as nvarchar(256), @Units as decimal, @Account as INT) AS
 BEGIN
 	DECLARE @MemberID INT
